requirements.txt
flask==3.0.0
pandas==2.2.1
python-docx==1.0.1
Werkzeug==3.0.1
gunicorn==21.2.0
docxtpl==0.16.7
openai==1.3.0
mongoengine==0.27.0
pymongo[srv]==4.6.0
python-dotenv==1.0.0
flask-cors==4.0.0
Pillow==10.1.0 